Refactored all data structures in application into shared module
