+++ src/app/manga-reader/manga-reader.component.ts
@@ -15,15 +15,9 @@ import { PageSplitOption } from '../_models/preferences/page-split-option';
 import { forkJoin } from 'rxjs';
 import { ToastrService } from 'ngx-toastr';
 import { KEY_CODES } from '../shared/_services/utility.service';
+import { CircularArray } from '../shared/data-structures/circular-array';
 
 const PREFETCH_PAGES = 3;
-
-// enum KEY_CODES {
-//   RIGHT_ARROW = 'ArrowRight',
-//   LEFT_ARROW = 'ArrowLeft',
-//   ESC_KEY = 'Escape',
-//   SPACE = ' '
-// }
 
 enum FITTING_OPTION {
   HEIGHT = 'full-height',
@@ -42,102 +36,6 @@ enum PAGING_DIRECTION {
   BACKWARDS = -1,
 }
 
-export class CircularArray<T> {
-  arr: T[];
-  currentIndex: number;
-
-  constructor(arr: T[], startIndex: number) {
-    this.arr = arr;
-    this.currentIndex = startIndex || 0;
-  }
-
-  next() {
-    const i = this.currentIndex;
-    const arr = this.arr;
-    this.currentIndex = i < arr.length - 1 ? i + 1 : 0;
-    return this.current();
-  }
-
-  prev() {
-    const i = this.currentIndex;
-    const arr = this.arr;
-    this.currentIndex = i > 0 ? i - 1 : arr.length - 1;
-    return this.current();
-  }
-
-  current() {
-    return this.arr[this.currentIndex];
-  }
-
-  peek(offset: number = 0) {
-    const i = this.currentIndex + 1 + offset;
-    const arr = this.arr;
-    const peekIndex = i < arr.length - 1 ? i + 1 : 0;
-    return this.arr[peekIndex];
-  }
-
-  size() {
-    return this.arr.length;
-  }
-
-  applyUntil(func: (item: T, index: number) => void, index?: number) {
-    index = index || this.currentIndex;
-    /// Applies a func against elements up until index. If index is 1 and size is 3, will apply on [2, 3, 0]
-    for (let offset = 1; offset < this.size(); offset++) {
-      const i = this.currentIndex + offset;
-      const arr = this.arr;
-      const peekIndex = i < arr.length ? i : 0;
-
-      if (peekIndex === index) {
-        break;
-      }
-
-      func(this.arr[peekIndex], peekIndex);
-    }
-
-  }
-
-  /// Applies a func against elements for X times. If limit is 1, size is 3, and index is 2. It will apply on [3]
-  applyFor(func: (item: T, index: number) => void, limit: number) {
-    for (let offset = 1; offset < limit; offset++) {
-      const i = this.currentIndex + offset;
-      const peekIndex = i < this.arr.length ? i : 0;
-
-      func(this.arr[peekIndex], peekIndex);
-    }
-
-  }
-
-
-}
-
-export class Queue<T> {
-  elements: T[];
-
-  constructor() {
-    this.elements = [];
-  }
-
-  enqueue(data: T) {
-    this.elements.push(data);
-  }
-
-  dequeue() {
-    return this.elements.shift();
-  }
-
-  isEmpty() {
-    return this.elements.length === 0;
-  }
-
-  peek() {
-    return !this.isEmpty() ? this.elements[0] : undefined;
-  }
-
-  length = () => {
-    return this.elements.length;
-  }
-}
 
 @Component({
   selector: 'app-manga-reader',
@@ -310,10 +208,6 @@ export class MangaReaderComponent implements OnInit, AfterViewInit, OnDestroy {
 
   toggleMenu() {
     this.menuOpen = !this.menuOpen;
-  }
-
-  getPageKey(pageNum: number) {
-    return `kavita-page-cache-${this.user.username}-${this.chapterId}--${pageNum}`;
   }
 
   isSplitLeftToRight() {
